requirements.txt
gradio
python-dotenv
langchain
pypdf
openai
chromadb
tiktoken
docx2txt
black